@@ -7,7 +7,6 @@ var HistoryStore = Flux.createStore({
         revisions: []
     },
 
-    loading: false,
 
     actions: {
         'history:load': 'loadHistory'
@@ -18,19 +17,14 @@ var HistoryStore = Flux.createStore({
     },
 
     loadHistory(params) {
-        this.loading = true;
         this.historyRepo.loadHistory(params).then((result) => {
             this.history = result;
-            this.loading = false;
             this.emitChange();
         });
     },
 
     getState() {
-        return {
-            value: this.history,
-            loading: this.loading
-        };
+        return this.history;
     }
 });
 
